# Muhash parallel reduce -- optimize U3072 mul when LHS = one