… consensus/benches/hash_benchmarks.rs